snowflake_blueprints/requirements.txt
snowflake-connector-python[pandas]
pandas==1.1.0
